Net Police
I have no actual expertise with Clojure, but read code well enough to be dangerous. Just looked through other implemented cards and wondered if this would be valid based on what I *think* I figured out. The only part I'm not sure about is if this would continually get re-evaluated or be set in stone based on Runner's link when rezzed.