Retain counters/credits on derezzed Corp cards
Per the Official FAQ (page 4), a derezzed Corp card needs to *retain* all hosted credits, cards, and counters. 

This is relevant now with the introduction of Test Ground, which is useful for reloading an Adonis or Eve Campaign with credits prior to them running out and self-trashing (and if you're really on the ball you'll re-rez for free with Breaker Bay Grid). 

Currently we're only keeping advancement tokens on things like the advanceable Weyland ice. What I *think* this will do is stop dissociating `;counter` no matter what, then only proceed to dissociate them if it's a Runner card (Imps, Gravediggers, et al) or if `keep-counter true` has been passed via the `derez` function.