Add partial implementation of Pheromones
Might as well toss in a partial for Pheromones to address #299. It isn't the whole enchilada since the CSS doesn't allow for a card that hosts both recurring credits *and* counters right now, but at least players can have a visual aid for tracking how many credits can be used in a turn.