When installing a card in a remote, don't remove that remote
fixes mtgred/netrunner#393 and mtgred/netrunner#156

When you installed an asset/agenda in a remote n that already had an
asset/agenda, that already installed card got trashed.

Then that server was removed from the game, and the remove server n+1
became server n, since the remotes are implicitly numbered.

Then the new asset/agenda was installed in server n, which was the wrong
server.