Make run-credit a separate UI-changeable property
Separate credit and run-credit in the UI. Run credits are only shown
in the UI when available. Automate spending run credits to satisfy credit
costs when possible on the backend as well.

This is mostly a change to support manually spending run credits
when we currently manually pay regular credits (e.g. pop-up or gargarin)
without having to worry about the wrong amount of credits returned
to the bank at the end of the run.