Improved card list scrolling performance
A lot of elements are being repainted when scrolling the card list.
This is taking a lot of time and the frame rate is restricted to around 20fps.
By moving the cards to a seperate layer, no more painting is needed when scrolling.
We are left with compositing only, and this is hardware accelerated on the gpu.
This boosts the performance to around 60fps.