Add installed-byname helper function and update in-play?
This function will look up whether a card with the given name is
currently installed. This is useful for checking if cards like Gene
Conditioning Shoppe are installed, as they are not targets during card
events or abilities.

in-play? used almost the same logic, so modfy it to avoid duplication.