Add functionality for Gene Conditioning Shoppe to genetics cards
Fairly strightforward change, basically we add an or condition to the
requirements for each card to trigger that will also trigger the
second time if GCS is installed.

Unfortunately, this technique will have to be used if new genetics
cards are introduced, as it's (AFAICT) impossible to implement this
behaviour on GCS itself.