Implement agenda Eden Fragment with manual ability
An automated implementation of Eden Fragment seemed difficult or
impossible to implement with the given structure for events related to
corp installs.  It also seemed difficult to handle the case where Eden
Fragment is scored after ICE has already been installed during the same
turn.

A manually triggered ability seems OK for an initial implementation.
There are no counters or limits to how many times the ability may be
triggered.  But the log (and the card) state clearly that it must be
used for the first piece of ICE installed that turn.