Fix for #948 (and some clean up)
When looking for card names in the text message, it's not enough to check
if the found card name does not end with ~ (the then used seperator for
the card text tokens), we also have to make sure that no ] (the then used
closing tag) follows; so we don't the regex to match

  Alpha

in this text

  Someone played [Hellion Alpha Test~01110] or something.

because it's already a card text token.

I also replaced the used [ ] and ~ characters with some characters you
don't find on a keyboard and extracted them into a variable.