fix #561: dont show card text when img-url present
Previously we were rendering the card text and then putting the img over the top of this.
When the card text was long this would result in card text being visible at the bottom.
We now only show the card text if we are unable to create the img url for the card.