Add correct Offer You Can't Refuse card to Corp score area
There's some instances the Runner takes a little while to decide. In the
meanwhile, if some other cards go into Archives, the last card discarded
would be the one added to the Corp score area.

This change ensures the actual card that is played to trigger the Runner
decision is the one added to the Corp score area if he/she decides not
to run.