Rewrite docker-compose file for development
The old docker file didn't work, so I assume it's not used in production
anywhere. This one should allow new developers to get jnet up and
running with a single command. Use the compose.sh script, as
unfortunately there is a dependency on the npm container being set up
before everything else.