"Fixed" Jua. It will now work if the corp targets facedown cards BUT: - As implemented it will show the runner a choice of card names, so if   by chance the facedown card chosen is the same card as the faceup one,   the runner will have no way of knowing which one is the facedown one   (technically they're in reverse order of what's written in the system   message, but that's hardly user-friendly) - I've written a second method to let the runner click the card chosen   by corp, but if there are more than one facedown card, the runner has   afaik no way of knowing which one corp chose (except clicking them)   The runner can click their facedown in the order they want them to   disappear, in case they'd rather shuffle in the faceup over certain   facedown, but also not very user-friendly.
I tried adding "cardname (facedown)" by updating :title in the list of
options, and then finding the card without the updated title via :cid
afterwards, but this did not work, as the list did not use the :title
in the card. I tried understanding exactly how the prompt worked to see
why this would be, but am none the wiser.