Made corp-install event awaitable
As :corp-install now triggers with trigger-event-sync, we can
get the problem where (as with ASA) the event triggering triggers
a new :corp-install. As the first is not finished, it will not be
registered that :corp-install has already happened, and as
trigger-event-sync checks req when triggering the event, this will
cause problems. I solved it by just setting all cards that trigger
on first-event? to also be :once :per-turn

Exceptions:
Comet (as you would then not be able to use the click ability to play
the second event)
Genetics (as they sometimes can trigger twice)

I might change trigger-event-sync to do what trigger-event-simult
does and check requirements first, then we can register the event
as happened before triggering the effects.