Implement Sundew
The :runner-spent-click event happens before the run is initiated, so we
do not know which server the run is going to happen on. This prevents us
from fully handling this ability in :runner-spent-click.

To fix, add a new field in the register, :click-type, which we can use
to determine if the click was spent on a run event. If so, we then wait
for the :begin-run event to handle the ability.