Change draw to use move
Currently, `draw` directly moves the given number of cards directly from the
deck to the hand, merely changing their zone.  This makes it very hard to watch
for card movement, as you have to duplicate your efforts either by let-binding
an ability and assigning it to many events (easy to forget some) or by directly
writing out the ability on each event.

This changes `draw` to call `move` the given number of times, which is slower,
yes, but also handles all of the edge cases where a card's data isn't be set
correctly by any outside ability. And by using `move`, we get the benefit of
calling `card-move` for every card drawn, giving us the chance to react in an
ability.