Move Buffer Drive trigger target after discard finishes
This is more in keeping with the rules; Buffer Drive triggers after the
trash, so the card should be moved from the Heap to the Stack. This
means, for example, that Skorpios Defence Systems will always take
precedence over Buffer Drive, regardless of whose turn it is.

The hanging effect is accomplished by registering an event on
:card-moved for the target while the trash is resolving.  This should
fire once the trash finishes resolving, and the target is moved to the
discard pile; the event's effect then moves the target to the bottom of
the deck.