Add tests for bad Buffer Drive trigger
Under this implementation of Buffer Drive, the sequence Corp trashes a
card in Grip -> install Buffer Drive -> Runner trashes a card in Grip
causes Buffer Drive to fire. I believe this is because the we are
checking whether the trash event was first-event? on a per-side basis,
although this doesn't explain why the same sequence with inverted sides
works...