Radically simplify Aniccam tests
Previous tests checked a number of irrelevant pieces of game state. Now
we simply check if the card we expect to have drawn is in the Grip.
Since Aniccam doesn't manipulate the Stack or Heap, checking that cards
have been discarded correctly and so on should not be within the scope
of these tests.