Initial implementation of event locations
With this set up, instead of remembering to register and unregister events in
the many places, instead we handle all registering and unregistering of events
in `move-card` (and presumably perform an initial registering for cards in deck
and the identity). This allows us to put a `:location` property on events,
specifying where the event should be active. By defaulting to the obvious places
(as determined by `default-location`), we can elide most of the location
statements, instead only specifying when outside of the norm.

This will play nicely once we move to handling Event and Operations properly
(leaving them on the table until their effect is completed), and will allow for
moving away from card def properties such as `:move-zone` and `:access`, as
those should be happening at the same time as their `trigger-event` brethren.

Uses the Shards as the example.