Rewrite Sandburg, add events to pay-credits
Rewrites Sandburg to not rely on `watch`, which really shouldn't be used
anywhere in the engine. Changes the `deduct` calls in `pay-credits` to be `lose`
calls, which allows us to watch for credit changes in a generic way: `:X-gain`
and `:X-lose`. Removes references to `deduct` outside of `game.core`, as it's an
implementation-level function. Much better to use `gain` and `lose`, or the more
specific functions (`gain-bad-publicity`, etc).