Resize zoomed card when resizing log
Resize the zoomed card so that it fills the width of the resizable log.
This allows players to resize the log so that they can more easily read
the zoomed card text.

The downside is that the height of the log will now be shorter than it
was previously. Hopefully this is a reasonable tradeoff.

Additionally remove the css transition for the card zoom otherwise the
resize becomes laggy due to the delay occurring every time the 'resize'
event fires.