Fix Buffer Drive crashing when triggered ability prompt is clicked
This is due to prompt logic being unable to handle cases where the card
referred to in the propmt has moved zone.

The workaround is hardcode an overwrite of the card's zone to [:discard]
in the Buffer Drive card implementation. This is likely somewhat
fragile; it is unclear how the solution will behave if other effects
cause the relevant cards to move.