Move base define-card back to card_defs
This is to keep the card-defs namespace focused on holding and defining the card
definitions. game.core needs a specific implementation so we can give it one in
game.core. The same can be done in other namespaces now as needed. I'm not sure
if using the same name in other namespaces is good, but I'll punt on that issue
for now.