changed calls to if' with potential to print the whole state
When `game.utils-test/if'` is called with a function in the first
position of the assertion (like in `(if' (prompt-is-type? state side
:select))`) and that test fails, the sub-forms of the assertion finally
get evaluated in `game.utils-test/error-wrapper` when given to
`do-report`. If state is one of those sub-forms, the whole state appears
in the report.

If tests get run in the repl or with lein, the whole state gets printed
into the console or log. `lein test` even crashes with a Java
OutOfMemory error after a while.

To prevent this, for all of the occurences of that pattern, we test the
assertion indirectly.

In addidtion, the test if a prompt is shown in `play-run-event-impl` did
not return an exception when no prompt is shown (because `(get-in @state
[:runner :prompt])` returns an empty list, when no prompt is shown).