translations: Add data translation support
Localized data (card titles, text, etc) can be retrieved using the
tr-data function if the user has selected a language other than
English. If the selected language doesn't have a translation, the
default (English) data is retrieved.

Translated data (if available) is present in the card data under the
':localized' key in the following format for available languages:

{... :localized {:title "...", :text "...", ...}}