translations: Add card language option
The new card language option controls the language used for card
data (which currently influences card data displayed in the card
browser and card names displayed during games). The current language
option now controls the language for the UI rather than all language
settings.

This allows users to select different languages for the UI and card
names, which is particularly useful if localized names are not known
and/or commonly used.